Remove unused R_DrawRectangle Add R_DrawFill (0.3.2) Update egl.pkz to 0.3.2 version Add features that egl.pkz 0.3.2 require (clamps/t, flare, blood decal/particle tables, etc)

+++ renderer/r_typedefs.h
@@ -37,7 +37,7 @@ typedef uint16			meshFeatures_t;
 // Image subsystem
 //
 typedef char			texUnit_t;
-typedef uint16			texFlags_t;
+typedef uint32			texFlags_t;
 
 //
 // Material subsystem
